requirements.txt
addfips==0.3.1
colorama==0.4.4
dash==2.3.0
Flask_Caching==1.10.1
h5py==3.6.0
nltk==3.7
numpy==1.22.3
pandas==1.4.2
plotly==5.6.0
psycopg2==2.9.3
pyproj==3.3.0
pytest==7.1.2
python-dotenv==0.20.0
requests==2.27.1
scikit_learn==1.1.1
scipy==1.8.0
setuptools==62.2.0
stateplane==0.4.0
tqdm==4.62.3
us==2.0.2
